Use uint64_t for line count and column to be on the safe side.
And add opl_parse_changeset_id() function and use it.

+++ include/osmium/io/detail/opl_input_format.hpp
@@ -58,7 +58,7 @@ namespace osmium {
 
                 osmium::memory::Buffer m_buffer{1024*1024};
                 const char* m_data;
-                size_t m_line_count = 0;
+                uint64_t m_line_count = 0;
 
                 void maybe_flush() {
                     if (m_buffer.committed() > 800*1024) {
